+++ packages/commons/src/sqs/index.ts
@@ -1,3 +1,2 @@
 export * as SQS from "./sqs.js";
 export * from "./utils.js";
-export * from "./sqsErrors.js";
