+++ packages/jwt-audit-analytics-writer/test/jwAuditService.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it, vi, afterAll } from "vitest";
+import { SQS } from "pagopa-interop-kpi-commons";
 import {
-  SQS,
+  CommonErrorCodes,
   decodeSQSEventMessageError,
-  ErrorCodes,
-} from "pagopa-interop-kpi-commons";
-import { InternalError } from "pagopa-interop-kpi-models";
+  InternalError,
+} from "pagopa-interop-kpi-models";
 import { processMessage } from "../src/handlers/messageHandler.js";
 import { sqsMessagesMock } from "./utils.js";
 
@@ -50,7 +50,7 @@ describe("JWT Audit Service tests", () => {
         await processMessage(mockJwtAuditService)(invalidMessage);
       } catch (error) {
         expect(error).toBeInstanceOf(InternalError);
-        expect((error as InternalError<ErrorCodes>).code).toBe(
+        expect((error as InternalError<CommonErrorCodes>).code).toBe(
           "decodeSQSEventMessageError"
         );
       }
